mysql.ej1/requirements.txt
mysqlclient==1.3.9m
mysql-connector-python-rf